fix various compiler warnings

+++ dashboard/src/drive.c
@@ -133,7 +133,7 @@ void drive_gear_update(uint32_t val) {
 void drive_speed_update(uint32_t val) {
     char str[8];
     val /= 100;
-    sprintf(str, "%3d", val);
+    sprintf(str, "%3u", (unsigned int)val);
     scr_draw_text(SCR_TEXT_ADDR(0, 17, 1), str);
 }
 
@@ -179,6 +179,6 @@ void drive_wb_radio_update(uint32_t val) {
 
 void drive_wb_traction_knob_update(uint32_t val) {
     char str[10];
-    sprintf(str, "%02d", val);
+    sprintf(str, "%02u", (unsigned int)val);
     scr_draw_text(SCR_TEXT_ADDR(0, 20, 7), str);
 }
